made swiper for tasks into a component
